Update WeatherForecast.cs
Burro

+++ APITempoDIO/WeatherForecast.cs
@@ -8,7 +8,7 @@ namespace APITempoDIO
 
         public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
 
-        public int TemperaturK => 273.16 + TemperatureC;
+        public double TemperaturK => 273.16 + TemperatureC;
 
         public string? Summary { get; set; }
     }
